escape ampersands properly in text
This ended up being a fairly large refactoring. Now I'm explicitly
distinguishing the strings that represent text from the strings that
represent HTML. This sort of thing is notoriously error-prone, and if
this was a framework for other people's consumption I'd want to make
this distinction explicit in the type system, but this should be good
enough for now.